Added support for circle segments in the 'Circle' shape.
Added the 'angle' and 'angleDeg' attributes to the 'Circle'
shape. The attribute defines the sweep angle for the circle
segment. The circle segment will hence sweep between [0-'angle']
radians (or degrees).

The idea is to define the size of the circle segment with the
'angle' attribute and then, if desired, rotate it with already
existing 'rotation' attribute.

Also added a unit test for the new functionality: 'SHAPE - add
circle segment using Circle'.